group chat

+++ external/emodel/chatmodel/chatpojo/group_item_mod_gen.go
@@ -17,7 +17,7 @@ type GroupItem struct {
 	MebNickname sql.NullString `db:"meb_nickname"` // 昵称
 	Remark      sql.NullString `db:"remarK"`       // 备注
 	ForbidenDur int32          `db:"forbiden_dur"` // 禁言时长(sec)
-	ForbiddenAt int8           `db:"forbidden_at"` // 禁言时间:[-1=永久,0=未禁言,other=tsMills]
+	ForbiddenAt int64          `db:"forbidden_at"` // 禁言时间:[-1=永久,0=未禁言,other=tsSec]
 	Cts         int64          `db:"cts"`          // 创建时间戳(ms)
 	Uts         sql.NullInt64  `db:"uts"`          // 更新时间戳(ms)
 }
